+++ xhci.device/src/xhci/xhci-udev.c
@@ -7,6 +7,7 @@
 #include <proto/exec.h>
 #endif
 
+#include <exec/errors.h>
 #include <devices/hcd_api.h>
 
 #include <xhci/xhci.h>
@@ -19,6 +20,8 @@
 #include <xhci/xhci-udev.h>
 #include <xhci/xhci-ring.h>
 #include <xhci/xhci-context.h>
+#include <xhci/xhci-lpm.h>
+#include <xhci/xhci-hub.h>
 
 #include <device.h>
 #include <debug.h>
@@ -36,13 +39,6 @@
 #undef KprintfH
 #define KprintfH(fmt, ...) PrintPistorm("[xhci-udev] %s: " fmt, __func__, ##__VA_ARGS__)
 #endif
-
-static void xhci_udev_parse_control_message(struct usb_device *udev, struct USBIORequest *io);
-static void xhci_udev_translate_hub_descriptor_request(struct usb_device *udev, struct USBIORequest *io);
-static struct usb_device *xhci_udev_find_child_on_port(struct usb_device *hub, u32 port);
-static BOOL xhci_udev_suspend_device(struct usb_device *udev, u8 root_port, struct USBIORequest *deferred_req);
-static void xhci_udev_do_set_config(struct usb_device *udev, struct USBIORequest *req);
-static void xhci_udev_op_clear(struct usb_device *udev);
 
 /* Move udev to a new virtual address.  The sole writer of the migration, so
  * the invariant map[addr] == udev <=> udev->virtual_address == addr lives in
@@ -225,49 +221,6 @@ static void xhci_udev_patch_endpoint_address(struct usb_device *udev, struct USB
              (ULONG)wIndex, (ULONG)fixed, (ULONG)epnum);
 }
 
-static BOOL xhci_udev_filter_emulated_hub_ctrl_request(struct usb_device *udev, struct USBIORequest *io)
-{
-    if (!udev || !io || !udev->ss_hub_emulation)
-        return FALSE;
-
-    struct USBSetupPacket *setup = &io->setup;
-    if (setup->bRequest != USB_REQ_CLEAR_FEATURE && setup->bRequest != USB_REQ_SET_FEATURE)
-        return FALSE;
-
-    if ((setup->bmRequestType & (USB_TYPE_MASK | USB_RECIP_MASK)) != (USB_TYPE_CLASS | USB_RECIP_OTHER))
-        return FALSE;
-
-    const u16 wValue = le16(setup->wValue);
-    const u8 portNo = le16(setup->wIndex) & 0xFFU;
-    switch (wValue)
-    {
-    case USB_PORT_FEAT_SUSPEND:
-    {
-        const u8 link_state = (setup->bRequest == USB_REQ_CLEAR_FEATURE) ? 0 : 3;
-        setup->wValue = le16(USB_PORT_FEAT_LINK_STATE);
-        setup->wIndex = le16(portNo | (link_state << 8));
-        return FALSE;
-    }
-
-    // these 3 are only for CLEAR_FEATURE
-    case USB_PORT_FEAT_ENABLE:
-        /* Can't disable USB 3.x port */
-    case USB_PORT_FEAT_C_ENABLE: // this is only used for clear feature
-        io->actual_length = 0;
-        io->req.io_Error = ERR_NO_ERROR;
-        if (!(io->req.io_Flags & IOF_QUICK))
-            ReplyMsg((struct Message *)io);
-        return TRUE;
-
-    case USB_PORT_FEAT_C_SUSPEND: // this is only used for clear feature
-        setup->wValue = le16(USB_SS_PORT_FEAT_C_LINK_STATE);
-        return FALSE;
-
-    default:
-        return FALSE;
-    }
-}
-
 s8 xhci_udev_send_ctrl(struct usb_device *udev, struct USBIORequest *io)
 {
     if (!udev || !io)
@@ -402,201 +355,6 @@ static BOOL xhci_udev_fetch_bos(struct usb_device *udev)
     return TRUE;
 }
 
-static s8 xhci_udev_send_ctrl_first(struct usb_device *udev, struct USBIORequest *io, u32 timeout_ms)
-{
-    /* Work around class drivers that omit the direction bit in endpoint-recipient requests (e.g., UAC1 SET_CUR). */
-    xhci_udev_patch_endpoint_address(udev, io);
-
-    KprintfH("bmReqType=%02lx bReq=%02lx wValue=%04lx wIndex=%04lx wLength=%04lx\n",
-             (ULONG)io->setup.bmRequestType,
-             (ULONG)io->setup.bRequest,
-             le16(io->setup.wValue),
-             le16(io->setup.wIndex),
-             le16(io->setup.wLength));
-
-    /* Translate USB 2.0 hub requests to SS format for SS hubs */
-    xhci_udev_translate_hub_descriptor_request(udev, io);
-
-    if (xhci_udev_filter_emulated_hub_ctrl_request(udev, io))
-        return ERR_NO_ERROR;
-
-    struct xhci_ctrl *ctrl = udev->controller;
-    if (io->virtual_address == xhci_roothub_get_address(ctrl->root_hub))
-    {
-        xhci_roothub_submit_ctrl_request(ctrl->root_hub, io);
-        xhci_udev_parse_control_message(udev, io);
-        if (io->req.io_Error != ERR_NO_ERROR)
-            return io->req.io_Error;
-        if (!(io->req.io_Flags & IOF_QUICK))
-            ReplyMsg((struct Message *)io);
-        return ERR_NO_ERROR;
-    }
-
-    struct USBSetupPacket *setup = &io->setup;
-    if (setup->bRequest == USB_REQ_SET_ADDRESS && (setup->bmRequestType & USB_TYPE_MASK) == USB_TYPE_STANDARD)
-    {
-        xhci_address_device(udev, io);
-        return ERR_NO_ERROR;
-    }
-
-    if (setup->bRequest == USB_REQ_SET_CONFIGURATION &&
-        (setup->bmRequestType & USB_TYPE_MASK) == USB_TYPE_STANDARD)
-    {
-        /* Stash the request as UDEV_OP_CONFIGURE, then optionally pre-fetch a
-         * descriptor on EP0 before configuring; the prefetch completion
-         * resumes the deferred SET_CONFIGURATION
-         * (xhci_udev_run_pending_set_config).
-         *   - Hubs: GET_DESCRIPTOR(hub) so xhci_set_configuration() can program
-         *     correct Number of Ports / TT Think Time into the slot context.
-         *   - HS/SS: GET_DESCRIPTOR(BOS) for U1/U2 exit latencies (USB3) or BESL
-         *     (USB2 LPM), needed for Max Exit Latency.  Non-fatal.
-         * EP0 is live (set up during ADDRESS_DEVICE).  A prefetch that can't be
-         * submitted leaves the stash intact, so we just run the config now. */
-        if (!xhci_udev_op_begin(udev, UDEV_OP_CONFIGURE, io))
-        {
-            /* op slot busy (unexpected): configure without the prefetch */
-            xhci_udev_do_set_config(udev, io);
-            return ERR_NO_ERROR;
-        }
-
-        if (udev->is_hub)
-        {
-            if (!udev->hub_desc_fetched && xhci_udev_fetch_hub_descriptor(udev))
-                return ERR_NO_ERROR;
-        }
-        else if (udev->speed >= USB_SPEED_HIGH)
-        {
-            if (!udev->bos_fetched && xhci_udev_fetch_bos(udev))
-                return ERR_NO_ERROR;
-        }
-
-        /* LS/FS, or the prefetch wasn't submitted: configure immediately. */
-        xhci_udev_run_pending_set_config(udev);
-        return ERR_NO_ERROR;
-    }
-
-    /* CLEAR_FEATURE(ENDPOINT_HALT) from the stack: keep the controller-side
-     * endpoint state in sync with the device (mirror Linux
-     * xhci_endpoint_reset).  If our own STALL recovery already reset the
-     * endpoint and sent the device-level clear-halt, answer without a
-     * duplicate wire request; if the endpoint is still halted (the stack
-     * noticed before our event handling), run our recovery - it ends with
-     * the internal clear-halt.  An unsolicited clear on a healthy endpoint
-     * (e.g. mass-storage reset recovery) is forwarded as-is: the device
-     * resets its data toggle/sequence, and a stale xHC sequence recovers
-     * through the normal STALL path on the next transfer. */
-    if (setup->bmRequestType == (USB_DIR_OUT | USB_TYPE_STANDARD | USB_RECIP_ENDPOINT) &&
-        setup->bRequest == USB_REQ_CLEAR_FEATURE &&
-        le16(setup->wValue) == USB_ENDPOINT_HALT &&
-        udev->slot_id != 0)
-    {
-        u8 ep_addr = (u8)(le16(setup->wIndex) & 0xffU);
-        u8 halt_ep_index = xhci_ep_index_from_parts(ep_addr & 0x0fU,
-                                                    (ep_addr & USB_DIR_IN) ? DIRECTION_IN : DIRECTION_OUT);
-        struct ep_context *halt_ep_ctx = (halt_ep_index != 0)
-                                             ? xhci_ep_get_context_for_index(udev, halt_ep_index)
-                                             : NULL;
-        if (halt_ep_ctx)
-        {
-            BOOL synced = xhci_ep_consume_halt_synced(halt_ep_ctx);
-            if (!synced && xhci_read_hw_ep_state(udev, halt_ep_index) == EP_STATE_HALTED)
-            {
-                KprintfH("stack clear-halt on halted EP %lu: running recovery\n", (ULONG)halt_ep_index);
-                xhci_reset_ep(udev, halt_ep_index);
-                synced = TRUE;
-            }
-            if (synced)
-            {
-                io->req.io_Error = ERR_NO_ERROR;
-                io->actual_length = 0;
-                if (!(io->req.io_Flags & IOF_QUICK))
-                    ReplyMsg((struct Message *)io);
-                return ERR_NO_ERROR;
-            }
-        }
-    }
-
-    /* SetPortFeature(SUSPEND) to an external hub: stop the attached child's
-     * endpoint rings first (xHCI 4.15.1); the request is forwarded to the hub
-     * once the stops complete (xhci_udev_suspend_finish). */
-    if (udev->is_hub &&
-        setup->bmRequestType == (USB_DIR_OUT | USB_RT_PORT) &&
-        setup->bRequest == USB_REQ_SET_FEATURE &&
-        le16(setup->wValue) == USB_PORT_FEAT_SUSPEND)
-    {
-        struct usb_device *child = xhci_udev_find_child_on_port(udev, le16(setup->wIndex) & 0xffU);
-        if (child && xhci_udev_suspend_device(child, 0, io))
-            return ERR_NO_ERROR;
-    }
-
-    /* If we don't have a slot yet, enable one and allocate Virt Dev */
-    if (udev->slot_id == 0 && udev->virtual_address == 0)
-    {
-        // this will store the req and submit it once addressed
-        xhci_address_device(udev, io);
-        return ERR_NO_ERROR;
-    }
-
-    s8 ret = xhci_ring_enqueue_td(udev, io, timeout_ms, FALSE);
-    return ret;
-}
-
-s8 xhci_udev_send(struct USBIORequest *req)
-{
-    struct XHCIUnit *unit = (struct XHCIUnit *)req->req.io_Unit;
-    if (!unit)
-    {
-        Kprintf("missing unit pointer (cmd=%lu, req=%lx, devaddr=%lu)\n",
-                (ULONG)req->req.io_Command, (ULONG)req, (ULONG)req->virtual_address);
-        return ERR_BAD_PARAMETERS;
-    }
-
-    struct usb_device *udev = xhci_udev_get(unit, req->virtual_address);
-    if (!udev)
-    {
-        KprintfH("Device does not exist for addr %lu\n", (ULONG)req->virtual_address);
-        return ERR_TIMEOUT;
-    }
-
-    u32 timeout_ms = 0;
-    if ((req->flags & DRIVER_FLAG_TIMEOUT_DEFINED))
-        timeout_ms = req->timeout;
-
-    KprintfH("dev=%lx addr=%lu slot=%lu ep=%lu dir=%s len=%lu flags=%lx tmo=%lu\n",
-             (ULONG)udev, (ULONG)udev->virtual_address, (ULONG)udev->slot_id,
-             (ULONG)(req->endpoint & 0x0F), (req->direction == DIRECTION_IN) ? "IN" : "OUT",
-             (ULONG)req->data_buffer_length, (ULONG)req->flags,
-             (ULONG)timeout_ms);
-
-    switch (req->req.io_Command)
-    {
-    case CMD_REQUEST_CONTROL:
-        return xhci_udev_send_ctrl_first(udev, req, timeout_ms);
-    case CMD_REQUEST_ISOCHRONOUS:
-    case CMD_REQUEST_BULK:
-        return xhci_ring_enqueue_td(udev, req, timeout_ms, FALSE);
-    case CMD_REQUEST_INTERRUPT:
-    {
-        struct xhci_ctrl *ctrl = unit->xhci_ctrl;
-        if (udev->virtual_address == xhci_roothub_get_address(ctrl->root_hub))
-        {
-            s8 result = xhci_roothub_submit_int_request(ctrl->root_hub, req);
-            return result;
-        }
-
-        return xhci_ring_enqueue_td(udev, req, timeout_ms, FALSE);
-    }
-    default:
-        Kprintf("unsupported command %lu (req=%lx, devaddr=%lu, endpoint=%lu, flags=0x%lx)\n",
-                (ULONG)req->req.io_Command,
-                (ULONG)req,
-                (ULONG)req->virtual_address,
-                (ULONG)req->endpoint,
-                (ULONG)req->flags);
-        return ERR_BAD_PARAMETERS;
-    }
-}
-
 /* Hooks for responding to requests for lower layer */
 void xhci_udev_io_reply_failed(struct xhci_ctrl *ctrl, struct USBIORequest *io, s8 err)
 {
@@ -616,10 +374,20 @@ void xhci_udev_io_reply_failed(struct xhci_ctrl *ctrl, struct USBIORequest *io, 
         {
             /* Internal EP0 requests are fire-and-forget; without this line a
              * device rejecting e.g. SET_FEATURE(U1/U2_ENABLE) or SET_SEL would
-             * be invisible. */
-            Kprintf("internal EP0 request failed: addr %lu bmReqType=%02lx bReq=%02lx wValue=%lu err=%ld\n",
-                    (ULONG)io->virtual_address, (ULONG)io->setup.bmRequestType,
-                    (ULONG)io->setup.bRequest, (ULONG)le16(io->setup.wValue), (LONG)err);
+             * be invisible.  But "device gone" (ERR_TIMEOUT) and "transfer
+             * cancelled" (IOERR_ABORTED) are the expected outcome when a hub or
+             * device is torn down: in-flight clear-halt / CLEAR_TT_BUFFER
+             * recovery races the physical removal and floods the log.  Keep
+             * those at debug level; a genuine device-level reject (STALL /
+             * Request Error) never reports timeout/aborted, so it stays loud. */
+            if (err == ERR_TIMEOUT || err == IOERR_ABORTED)
+                KprintfH("internal EP0 request retired (device gone): addr %lu bmReqType=%02lx bReq=%02lx wValue=%lu err=%ld\n",
+                         (ULONG)io->virtual_address, (ULONG)io->setup.bmRequestType,
+                         (ULONG)io->setup.bRequest, (ULONG)le16(io->setup.wValue), (LONG)err);
+            else
+                Kprintf("internal EP0 request failed: addr %lu bmReqType=%02lx bReq=%02lx wValue=%lu err=%ld\n",
+                        (ULONG)io->virtual_address, (ULONG)io->setup.bmRequestType,
+                        (ULONG)io->setup.bRequest, (ULONG)le16(io->setup.wValue), (LONG)err);
 
             /* A rejected SET_SEL ends the LPM sequence (no device-initiated
              * U1/U2); free the op slot. */
@@ -673,6 +441,16 @@ static void xhci_udev_handle_hub_prefetch(struct usb_device *udev, struct USBIOR
     xhci_udev_run_pending_set_config(udev);
 }
 
+/* Reset the per-device multi-step operation slot to idle. */
+static void xhci_udev_op_clear(struct usb_device *udev)
+{
+    udev->op.op = UDEV_OP_NONE;
+    udev->op.step = 0;
+    udev->op.waits = 0;
+    udev->op.arg = 0;
+    udev->op.stash = NULL;
+}
+
 /* Program the configuration and issue CONFIG_EP for req (the stashed
  * SET_CONFIGURATION travels onward as the command's completion request). */
 static void xhci_udev_do_set_config(struct usb_device *udev, struct USBIORequest *req)
@@ -710,22 +488,73 @@ void xhci_udev_run_pending_set_config(struct usb_device *udev)
     xhci_udev_do_set_config(udev, orig_req);
 }
 
-/**
- * xhci_udev_handle_bos_prefetch - BOS descriptor pre-fetch completion handler.
- *
- * Phase 1 (data_buffer_length == sizeof(struct usb_bos_descriptor)):
- *   Parse wTotalLength from the 5-byte BOS header and issue the full-BOS
- *   phase-2 fetch.  If the phase-2 ring submission succeeds we return without
- *   touching the UDEV_OP_CONFIGURE stash; otherwise we fall through and run
- *   SET_CONFIGURATION immediately with MEL = 0.
- *
- * Phase 2 (data_buffer_length > sizeof(struct usb_bos_descriptor)):
- *   Walk every Device Capability descriptor and populate the USB2/USB3
- *   exit-latency fields in udev, then run SET_CONFIGURATION.
- *
- * The io->data_buffer is always freed here.  The io struct itself is freed
- * by the caller (xhci_udev_io_reply_data) after we return.
- */
+/* Phase 1 of the BOS pre-fetch: validate the 5-byte header and issue the
+ * full-length BOS fetch.  Returns TRUE if phase 2 was submitted (the
+ * UDEV_OP_CONFIGURE stash stays put; the caller must not run set-config yet);
+ * FALSE if it should give up and run the deferred SET_CONFIGURATION now.  On
+ * submit or a post-free give-up *pbuf is cleared so the caller does not double
+ * free; a header-validation give-up leaves *pbuf for the caller to free. */
+static BOOL xhci_udev_bos_prefetch_phase1(struct usb_device *udev, u8 **pbuf)
+{
+    struct xhci_ctrl *ctrl = udev->controller;
+    u8 *buf = *pbuf;
+
+    KprintfH("HUB descriptor pre-fetch phase 1: total length field is %lu\n",
+             (ULONG)le16(((struct usb_bos_descriptor *)buf)->wTotalLength));
+    const struct usb_bos_descriptor *hdr = (const struct usb_bos_descriptor *)buf;
+    if (hdr->bDescriptorType != USB_DT_BOS || hdr->bNumDeviceCaps == 0)
+        return FALSE;
+
+    u16 total = le16(hdr->wTotalLength);
+    if (total <= (u16)sizeof(struct usb_bos_descriptor) || total > 512u)
+    {
+        Kprintf("BOS wTotalLength %u out of range\n", (unsigned)total);
+        return FALSE;
+    }
+
+    dma_free(ctrl->dmaPool, buf);
+    *pbuf = NULL;
+
+    u8 *buf2 = dma_alloc(ctrl->dmaPool, DMA_ALIGN_MIN, (u32)total);
+    struct USBIORequest *io2 = pool_zalloc(ctrl->metaPool, sizeof(*io2));
+    if (!io2 || !buf2)
+    {
+        Kprintf("xhci_udev_handle_bos_prefetch: phase-2 alloc failed\n");
+        if (buf2)
+            dma_free(ctrl->dmaPool, buf2);
+        if (io2)
+            pool_free(ctrl->metaPool, io2);
+        return FALSE;
+    }
+
+    io2->req.io_Command = CMD_REQUEST_CONTROL;
+    io2->req.io_Flags = IOF_QUICK;
+    io2->driver_private_flags = REQ_INTERNAL | REQ_ENQUEUED | REQ_BOS_FETCH;
+
+    io2->setup.bmRequestType = USB_DIR_IN | USB_TYPE_STANDARD | USB_RECIP_DEVICE;
+    io2->setup.bRequest = USB_REQ_GET_DESCRIPTOR;
+    io2->setup.wValue = le16((u16)(USB_DT_BOS << 8));
+    io2->setup.wIndex = 0;
+    io2->setup.wLength = le16(total);
+
+    io2->virtual_address = udev->virtual_address;
+    io2->data_buffer = buf2;
+    io2->data_buffer_length = (u32)total;
+    io2->direction = DIRECTION_IN;
+
+    s8 ring_ret = xhci_ring_enqueue_td(udev, io2, 1000, FALSE);
+    if (ring_ret != ERR_NO_ERROR)
+    {
+        Kprintf("xhci_udev_handle_bos_prefetch: phase-2 enqueue failed (%ld)\n", (LONG)ring_ret);
+        dma_free(ctrl->dmaPool, buf2);
+        pool_free(ctrl->metaPool, io2);
+        return FALSE;
+    }
+
+    /* Phase 2 in flight; the UDEV_OP_CONFIGURE stash stays put */
+    return TRUE;
+}
+
 static void xhci_udev_handle_bos_prefetch(struct usb_device *udev, struct USBIORequest *io)
 {
     struct xhci_ctrl *ctrl = udev->controller;
@@ -739,141 +568,16 @@ static void xhci_udev_handle_bos_prefetch(struct usb_device *udev, struct USBIOR
 
     if (io->data_buffer_length == (u32)sizeof(struct usb_bos_descriptor))
     {
-        /* Phase 1: validate header and issue the full-BOS fetch */
-        KprintfH("HUB descriptor pre-fetch phase 1: total length field is %lu\n",
-                 (ULONG)le16(((struct usb_bos_descriptor *)buf)->wTotalLength));
-        const struct usb_bos_descriptor *hdr = (const struct usb_bos_descriptor *)buf;
-        if (hdr->bDescriptorType != USB_DT_BOS || hdr->bNumDeviceCaps == 0)
-            goto run_set_config;
-
-        u16 total = le16(hdr->wTotalLength);
-        if (total <= (u16)sizeof(struct usb_bos_descriptor) || total > 512u)
-        {
-            Kprintf("BOS wTotalLength %u out of range\n", (unsigned)total);
-            goto run_set_config;
-        }
-
-        dma_free(ctrl->dmaPool, buf);
-        buf = NULL;
-
-        u8 *buf2 = dma_alloc(ctrl->dmaPool, DMA_ALIGN_MIN, (u32)total);
-        struct USBIORequest *io2 = pool_zalloc(ctrl->metaPool, sizeof(*io2));
-        if (!io2 || !buf2)
-        {
-            Kprintf("xhci_udev_handle_bos_prefetch: phase-2 alloc failed\n");
-            if (buf2)
-                dma_free(ctrl->dmaPool, buf2);
-            if (io2)
-                pool_free(ctrl->metaPool, io2);
-            goto run_set_config;
-        }
-
-        io2->req.io_Command = CMD_REQUEST_CONTROL;
-        io2->req.io_Flags = IOF_QUICK;
-        io2->driver_private_flags = REQ_INTERNAL | REQ_ENQUEUED | REQ_BOS_FETCH;
-
-        io2->setup.bmRequestType = USB_DIR_IN | USB_TYPE_STANDARD | USB_RECIP_DEVICE;
-        io2->setup.bRequest = USB_REQ_GET_DESCRIPTOR;
-        io2->setup.wValue = le16((u16)(USB_DT_BOS << 8));
-        io2->setup.wIndex = 0;
-        io2->setup.wLength = le16(total);
-
-        io2->virtual_address = udev->virtual_address;
-        io2->data_buffer = buf2;
-        io2->data_buffer_length = (u32)total;
-        io2->direction = DIRECTION_IN;
-
-        s8 ring_ret = xhci_ring_enqueue_td(udev, io2, 1000, FALSE);
-        if (ring_ret != ERR_NO_ERROR)
-        {
-            Kprintf("xhci_udev_handle_bos_prefetch: phase-2 enqueue failed (%ld)\n", (LONG)ring_ret);
-            dma_free(ctrl->dmaPool, buf2);
-            pool_free(ctrl->metaPool, io2);
-            goto run_set_config;
-        }
-
-        /* Phase 2 in flight; the UDEV_OP_CONFIGURE stash stays put */
-        return;
+        /* Phase 1: validate header and issue the full-BOS fetch. */
+        if (xhci_udev_bos_prefetch_phase1(udev, &buf))
+            return; /* phase 2 in flight */
     }
     else
     {
-        /* Phase 2: parse all Device Capability descriptors */
-        const struct usb_bos_descriptor *hdr = (const struct usb_bos_descriptor *)buf;
-        if (hdr->bDescriptorType != USB_DT_BOS)
-            goto run_set_config;
-
-        KprintfH("BOS descriptor total length is %lu, num device caps is %lu\n",
-                 (ULONG)le16(hdr->wTotalLength), (ULONG)hdr->bNumDeviceCaps);
-        const u8 *cursor = buf + sizeof(struct usb_bos_descriptor);
-        const u8 *end = buf + (u32)le16(hdr->wTotalLength);
-        if (end > buf + io->data_buffer_length)
-            end = buf + io->data_buffer_length;
-
-        while (cursor + 3u <= end)
-        {
-            u8 caplen = cursor[0];
-            u8 dtype = cursor[1];
-            u8 captype = cursor[2];
-            if (caplen < 3u || cursor + caplen > end)
-                break;
-
-            if (dtype == USB_DT_DEVICE_CAPABILITY)
-            {
-                if (captype == USB_CAP_DESC_USB20_EXTENSION &&
-                    caplen >= (u8)sizeof(struct usb_2_0_extension_capability_descriptor))
-                {
-                    const struct usb_2_0_extension_capability_descriptor *ext =
-                        (const struct usb_2_0_extension_capability_descriptor *)cursor;
-                    u32 att = le32(ext->bmAttributes);
-                    udev->lpm_capable = (att & USB_20_EXTENSION_ATT_LINK_POWER_MANAGEMENT) ? TRUE : FALSE;
-                    udev->besl_supported = (att & USB_20_EXTENSION_ATT_BESL_SUPPORTED) ? TRUE : FALSE;
-                    udev->besl_baseline_valid = (att & USB_20_EXTENSION_ATT_BESL_BASELINE_VALID) ? TRUE : FALSE;
-                    udev->besl_deep_valid = (att & USB_20_EXTENSION_ATT_BESL_DEEP_VALID) ? TRUE : FALSE;
-                    if (udev->besl_baseline_valid)
-                        udev->besl_baseline = (u8)USB_20_EXTENSION_ATT_BESL_BASELINE(att);
-                    if (udev->besl_deep_valid)
-                        udev->besl_deep = (u8)USB_20_EXTENSION_ATT_BESL_DEEP(att);
-                    KprintfH("USB2 Ext Cap: lpm=%ld besl=%ld baseline=%lu(v=%ld) deep=%lu(v=%ld)\n",
-                             (LONG)udev->lpm_capable, (LONG)udev->besl_supported,
-                             (ULONG)udev->besl_baseline, (LONG)udev->besl_baseline_valid,
-                             (ULONG)udev->besl_deep, (LONG)udev->besl_deep_valid);
-                }
-                else if (captype == USB_CAP_DESC_SS_USB_DEVICE &&
-                         caplen >= (u8)sizeof(struct usb_ss_device_capability_descriptor))
-                {
-                    const struct usb_ss_device_capability_descriptor *ss =
-                        (const struct usb_ss_device_capability_descriptor *)cursor;
-                    udev->u1_dev_exit_lat = ss->bU1DevExitLat;
-                    udev->u2_dev_exit_lat = le16(ss->wU2DevExitLat);
-                    udev->ltm_capable = (ss->bmAttributes & USB_SS_DEVICE_ATT_LATENCY_TOLERANCE_MESSAGES) != 0;
-                    KprintfH("SS Dev Cap: U1=%lu U2=%lu\n",
-                             (ULONG)udev->u1_dev_exit_lat, (ULONG)udev->u2_dev_exit_lat);
-                }
-            }
-
-            cursor += caplen;
-        }
-        /* Fall through to run_set_config */
-        KprintfH("Finished parsing BOS device capabilities for addr=%lu\n", (ULONG)udev->virtual_address);
-        KprintfH("Device LPM capability: %ld, BESL support: %ld, BESL baseline: %lu\n",
-                 (LONG)udev->lpm_capable, (LONG)udev->besl_supported, (ULONG)udev->besl_baseline);
-        KprintfH("Device U1 exit latency: %lu us, U2 exit latency: %lu us\n",
-                 (ULONG)udev->u1_dev_exit_lat, (ULONG)udev->u2_dev_exit_lat);
-
-        /* For SS devices, LPM capability comes from the SS Device Cap exit
-         * latencies (the USB2 Ext cap LPM bit only governs USB2 L1).  Mirror
-         * usb_device_supports_lpm(): non-zero exit latency and an LPM-capable
-         * path to the root hub.  (USB2 lpm_capable was set from the Ext cap.) */
-        if (udev->speed >= USB_SPEED_SUPER)
-            udev->lpm_capable = ((udev->u1_dev_exit_lat || udev->u2_dev_exit_lat) &&
-                                 (!udev->parent || udev->parent->parent == NULL ||
-                                  udev->parent->lpm_capable))
-                                    ? TRUE
-                                    : FALSE;
-
-        /* Now that exit latencies are known, compute the USB3 SEL/PEL/MEL
-         * parameters (no-op for non-SS devices). */
-        xhci_set_lpm_parameters(udev);
+        /* Phase 2: hand the full BOS off to the LPM module, which parses the
+         * Device Capability descriptors into udev's LPM fields and computes the
+         * SEL/PEL/MEL parameters. */
+        xhci_lpm_parse_bos_caps(udev, buf, io->data_buffer_length);
     }
 
 run_set_config:
@@ -901,7 +605,7 @@ static void xhci_udev_complete_internal(struct usb_device *udev, struct USBIOReq
     /* Hub descriptor pre-fetch completion: now run the full
      * SET_CONFIGURATION + CONFIG_EP with real hub data available.
      * On the success path xhci_udev_parse_control_message already ran, so
-     * handle_get_hub_descriptor cached ss_hub_desc + tt_think_time. */
+     * xhci_hub_handle_get_descriptor cached ss_hub_desc + tt_think_time. */
     if (io->driver_private_flags & REQ_HUB_DESC_FETCH)
         xhci_udev_handle_hub_prefetch(udev, io);
 
@@ -938,33 +642,10 @@ static void xhci_udev_complete_internal(struct usb_device *udev, struct USBIOReq
         pool_free(ctrl->metaPool, io);
 }
 
-void xhci_udev_io_reply_data(struct usb_device *udev, struct USBIORequest *io, s8 err, u32 actual)
-{
-    if (!io || !udev)
-        return;
-
-    io->actual_length = actual;
-    io->req.io_Error = err;
-
-    if (io->req.io_Command == CMD_REQUEST_CONTROL && err == ERR_NO_ERROR && io->endpoint == 0)
-        xhci_udev_parse_control_message(udev, io);
-
-    KprintfH("err=%ld actual=%lu\n", (LONG)err, (ULONG)actual);
-
-    /* Internal, reply-less requests (IOF_QUICK + magic tag) */
-    if (io->driver_private_flags & REQ_INTERNAL)
-    {
-        xhci_udev_complete_internal(udev, io);
-        return;
-    }
-
-    ReplyMsg((struct Message *)io);
-}
-
-static inline void xhci_udev_send_control_request(struct usb_device *udev, u8 ep_index,
-                                                  u8 bmRequestType, u8 bRequest,
-                                                  u16 wValue, u16 wIndex, u16 wLength,
-                                                  BOOL enqueue)
+void xhci_udev_send_control_request(struct usb_device *udev, u8 ep_index,
+                                    u8 bmRequestType, u8 bRequest,
+                                    u16 wValue, u16 wIndex, u16 wLength,
+                                    BOOL enqueue)
 {
     if (!udev || !udev->controller)
         return;
@@ -1007,146 +688,6 @@ inline static u8 xhci_ep_index_to_address(u8 ep_index)
     return (u8)(EP_INDEX_TO_ENDPOINT(ep_index) | ((ep_index & 0x1U) ? USB_DIR_OUT : USB_DIR_IN));
 }
 
-/* Send SET_SEL (USB 3.2 9.4.12) to inform the device of host/path exit latencies.
- * Returns TRUE if the transfer was submitted (and device-initiated LPM may be
- * enabled), FALSE if SEL/PEL exceed the field range or submission failed. */
-BOOL xhci_udev_send_set_sel(struct usb_device *udev)
-{
-    if (!udev || !udev->controller || udev->speed < USB_SPEED_SUPER || !udev->lpm_capable)
-        return FALSE;
-
-    struct xhci_ctrl *ctrl = udev->controller;
-
-    /* Convert ns -> us (round up) */
-    u32 u1_sel = (udev->u1_sel + 999u) / 1000u;
-    u32 u1_pel = (udev->u1_pel + 999u) / 1000u;
-    u32 u2_sel = (udev->u2_sel + 999u) / 1000u;
-    u32 u2_pel = (udev->u2_pel + 999u) / 1000u;
-
-    if (u1_sel > USB3_LPM_MAX_U1_SEL_PEL || u1_pel > USB3_LPM_MAX_U1_SEL_PEL ||
-        u2_sel > USB3_LPM_MAX_U2_SEL_PEL || u2_pel > USB3_LPM_MAX_U2_SEL_PEL)
-    {
-        Kprintf("SET_SEL skipped for addr %lu: SEL/PEL too large (u1 %lu/%lu u2 %lu/%lu us)\n",
-                (ULONG)udev->virtual_address, (ULONG)u1_sel, (ULONG)u1_pel,
-                (ULONG)u2_sel, (ULONG)u2_pel);
-        return FALSE;
-    }
-
-    u8 *buf = dma_alloc(ctrl->dmaPool, DMA_ALIGN_MIN, sizeof(struct usb_set_sel_req));
-    struct USBIORequest *io = pool_zalloc(ctrl->metaPool, sizeof(*io));
-    if (!io || !buf)
-    {
-        Kprintf("xhci_udev_send_set_sel: alloc failed\n");
-        if (buf)
-            dma_free(ctrl->dmaPool, buf);
-        if (io)
-            pool_free(ctrl->metaPool, io);
-        return FALSE;
-    }
-
-    struct usb_set_sel_req *sel = (struct usb_set_sel_req *)buf;
-    sel->u1_sel = (u8)u1_sel;
-    sel->u1_pel = (u8)u1_pel;
-    sel->u2_sel = le16((u16)u2_sel);
-    sel->u2_pel = le16((u16)u2_pel);
-
-    io->req.io_Command = CMD_REQUEST_CONTROL;
-    io->req.io_Flags = IOF_QUICK;
-    io->driver_private_flags = REQ_INTERNAL | REQ_ENQUEUED | REQ_SET_SEL;
-
-    io->setup.bmRequestType = USB_DIR_OUT | USB_TYPE_STANDARD | USB_RECIP_DEVICE;
-    io->setup.bRequest = USB_REQ_SET_SEL;
-    io->setup.wValue = 0;
-    io->setup.wIndex = 0;
-    io->setup.wLength = le16((u16)sizeof(struct usb_set_sel_req));
-
-    io->virtual_address = udev->virtual_address;
-    io->data_buffer = buf;
-    io->data_buffer_length = (u32)sizeof(struct usb_set_sel_req);
-    io->direction = DIRECTION_OUT;
-
-    s8 ring_ret = xhci_ring_enqueue_td(udev, io, 1000, FALSE);
-    if (ring_ret != ERR_NO_ERROR)
-    {
-        Kprintf("xhci_udev_send_set_sel: enqueue failed (%ld)\n", (LONG)ring_ret);
-        dma_free(ctrl->dmaPool, buf);
-        pool_free(ctrl->metaPool, io);
-        return FALSE;
-    }
-
-    KprintfH("SET_SEL addr %lu: u1 sel=%lu pel=%lu, u2 sel=%lu pel=%lu (us)\n",
-             (ULONG)udev->virtual_address, (ULONG)u1_sel, (ULONG)u1_pel,
-             (ULONG)u2_sel, (ULONG)u2_pel);
-    return TRUE;
-}
-
-/* Enable device-initiated U1 or U2 transitions via SET_FEATURE.  Fire-and-forget. */
-void xhci_udev_set_device_lpm(struct usb_device *udev, BOOL u2)
-{
-    u8 feature = u2 ? USB_DEVICE_U2_ENABLE : USB_DEVICE_U1_ENABLE;
-    xhci_udev_send_control_request(udev, 0,
-                                   USB_DIR_OUT | USB_TYPE_STANDARD | USB_RECIP_DEVICE,
-                                   USB_REQ_SET_FEATURE,
-                                   feature /* wValue */,
-                                   0 /* wIndex */,
-                                   0 /* wLength */,
-                                   FALSE /* send now */);
-    Kprintf("SET_FEATURE %s_ENABLE addr %lu\n", u2 ? "U2" : "U1", (ULONG)udev->virtual_address);
-}
-
-/* Enable device-initiated Latency Tolerance Messaging via SET_FEATURE.
- * Fire-and-forget; the xHC consumes the resulting LTM packets in hardware
- * (HCC_LTC) and uses the device's BELT for its U-state timing. */
-void xhci_udev_set_device_ltm(struct usb_device *udev)
-{
-    xhci_udev_send_control_request(udev, 0,
-                                   USB_DIR_OUT | USB_TYPE_STANDARD | USB_RECIP_DEVICE,
-                                   USB_REQ_SET_FEATURE,
-                                   USB_DEVICE_LTM_ENABLE /* wValue */,
-                                   0 /* wIndex */,
-                                   0 /* wLength */,
-                                   FALSE /* send now */);
-    Kprintf("SET_FEATURE LTM_ENABLE addr %lu\n", (ULONG)udev->virtual_address);
-}
-
-/* Program the U1/U2 inactivity timeout on the device's immediate parent-hub
- * downstream port.  A device directly on the root hub writes the root port's
- * PORTPMSC register; a device behind an external hub gets a
- * SetPortFeature(U1/U2_TIMEOUT) control transfer sent to that hub, which programs
- * its own downstream port (mirrors usb_set_lpm_timeout()).  timeout==0 disables
- * the state.  Fire-and-forget for the external-hub case. */
-void xhci_udev_set_port_lpm_timeout(struct usb_device *udev, BOOL u2, u16 timeout)
-{
-    if (!udev || !udev->controller)
-        return;
-
-    struct usb_device *parent = udev->parent;
-    if (!parent)
-        return;
-
-    if (parent->parent == NULL)
-    {
-        /* Direct root-hub child: write the root port PORTPMSC directly. */
-        xhci_roothub_set_usb3_port_timeout(udev->controller->root_hub,
-                                           udev->parent_port, u2, timeout);
-    }
-    else
-    {
-        /* Behind an external hub: ask the hub to program its downstream port.
-         * wValue = feature selector, wIndex = (timeout << 8) | port. */
-        u8 feature = u2 ? USB_SS_PORT_FEAT_U2_TIMEOUT : USB_SS_PORT_FEAT_U1_TIMEOUT;
-        u16 wIndex = (u16)(((u16)timeout << 8) | (u16)udev->parent_port);
-        xhci_udev_send_control_request(parent, 0,
-                                       USB_DIR_OUT | USB_RT_PORT,
-                                       USB_REQ_SET_FEATURE,
-                                       feature /* wValue */, wIndex,
-                                       0 /* wLength */, FALSE /* send now */);
-        Kprintf("SetPortFeature %s_TIMEOUT=%lu to hub addr %lu port %lu (for dev addr %lu)\n",
-                u2 ? "U2" : "U1", (ULONG)timeout, (ULONG)parent->virtual_address,
-                (ULONG)udev->parent_port, (ULONG)udev->virtual_address);
-    }
-}
-
 /* ---- Multi-step device operations ------------------------------------------
  * One operation in flight per device (struct udev_operation).  Cross-module
  * completion sites feed xhci_udev_op_advance(); op-specific steps live in its
@@ -1154,15 +695,6 @@ void xhci_udev_set_port_lpm_timeout(struct usb_device *udev, BOOL u2, u16 timeou
  * operation preempts it; everything else is mutually exclusive by
  * construction (and rejected with a log if that assumption ever breaks).
  */
-
-static void xhci_udev_op_clear(struct usb_device *udev)
-{
-    udev->op.op = UDEV_OP_NONE;
-    udev->op.step = 0;
-    udev->op.waits = 0;
-    udev->op.arg = 0;
-    udev->op.stash = NULL;
-}
 
 /* Abort the in-flight operation (which == UDEV_OP_NONE matches any), replying
  * the stashed request so the stack isn't left waiting. */
@@ -1444,272 +976,6 @@ s32 xhci_ep_type_for_index(struct usb_device *udev, u8 ep_index)
     return -1;
 }
 
-static void parse_config_descriptor(struct usb_device *udev, u8 *data, u16 len)
-{
-    if (len < 2)
-    {
-        KprintfH("too short, len=%lu\n", (ULONG)len);
-        return;
-    }
-
-    struct usb_config *conf = pool_zalloc(udev->controller->metaPool, sizeof(*conf));
-    if (!conf)
-    {
-        Kprintf("pool_zalloc failed\n");
-        return;
-    }
-
-    struct usb_config_descriptor *desc = (struct usb_config_descriptor *)data;
-    if (desc->bDescriptorType != USB_DT_CONFIG)
-    {
-        Kprintf("bad desc type %lu\n", (ULONG)desc->bDescriptorType);
-        goto error;
-    }
-
-    u16 total_len = le16(desc->wTotalLength);
-    if (len < total_len)
-    {
-        KprintfH("short buffer len=%lu total_len=%lu\n", (ULONG)len, (ULONG)total_len);
-        return;
-    }
-
-    u8 *cursor = data;
-    u8 *end = data + total_len;
-    if (cursor + desc->bLength > end)
-    {
-        Kprintf("bad desc length %lu\n", (ULONG)desc->bLength);
-        goto error;
-    }
-
-    CopyMem(desc, &conf->desc, sizeof(struct usb_config_descriptor));
-    cursor += desc->bLength;
-
-    KprintfH("wTotalLength=%lu bNumInterfaces=%lu bConfigurationValue=%lu iConfiguration=%lu bmAttributes=0x%02lx bMaxPower=%lu\n",
-             (ULONG)le16(desc->wTotalLength),
-             (ULONG)desc->bNumInterfaces,
-             (ULONG)desc->bConfigurationValue,
-             (LONG)desc->iConfiguration,
-             (LONG)desc->bmAttributes,
-             (LONG)desc->bMaxPower);
-
-    // in 3.x there are association descriptors here
-
-    int interface_map[USB_MAXINTERFACES];
-    for (int i = 0; i < USB_MAXINTERFACES; ++i)
-        interface_map[i] = -1;
-
-    conf->no_of_if = 0;
-    int if_index = 0;
-    int current_alt_index = -1;
-    struct usb_interface *current_if = NULL;
-    struct usb_interface_altsetting *current_alt = NULL;
-
-    while (cursor + 2 <= end)
-    {
-        u8 dlen = cursor[0];
-        u8 dtype = cursor[1];
-        if (dlen == 0)
-        {
-            Kprintf("zero length descriptor, aborting\n");
-            break;
-        }
-        if (cursor + dlen > end)
-        {
-            Kprintf("descriptor overruns buffer (type=%lu len=%lu)\n", (ULONG)dtype, (ULONG)dlen);
-            break;
-        }
-
-        switch (dtype)
-        {
-        case USB_DT_INTERFACE:
-        {
-            struct usb_interface_descriptor *ifd = (struct usb_interface_descriptor *)cursor;
-            u32 iface_number = ifd->bInterfaceNumber;
-            if (iface_number >= USB_MAXINTERFACES)
-            {
-                Kprintf("interface number %lu exceeds max %lu\n", (ULONG)iface_number, (ULONG)USB_MAXINTERFACES);
-                current_if = NULL;
-                current_alt = NULL;
-                current_alt_index = -1;
-                break;
-            }
-
-            if_index = interface_map[iface_number];
-            if (if_index < 0)
-            {
-                if_index = conf->no_of_if;
-                if (if_index >= USB_MAXINTERFACES)
-                {
-                    Kprintf("too many unique interfaces (%lu)\n", (ULONG)if_index);
-                    goto error;
-                }
-                interface_map[iface_number] = if_index;
-                current_if = &conf->if_desc[if_index];
-                mem_zero(current_if, sizeof(struct usb_interface));
-                current_if->interface_number = (u8)iface_number;
-                current_if->num_altsetting = 0;
-                current_if->active_altsetting = NULL;
-                conf->no_of_if++;
-            }
-            else
-            {
-                current_if = &conf->if_desc[if_index];
-            }
-
-            current_if->interface_number = (u8)iface_number;
-
-            if (current_if->num_altsetting >= USB_ALTSETTINGALLOC)
-            {
-                Kprintf("too many alternate settings (%lu) for interface %lu\n",
-                        (ULONG)current_if->num_altsetting, (ULONG)iface_number);
-                current_alt = NULL;
-                current_alt_index = -1;
-                break;
-            }
-
-            current_alt_index = current_if->num_altsetting++;
-            current_alt = &current_if->altsetting[current_alt_index];
-            mem_zero(current_alt, sizeof(struct usb_interface_altsetting));
-
-            CopyMem(ifd, &current_alt->desc, sizeof(struct usb_interface_descriptor));
-            current_alt->no_of_ep = 0;
-
-            KprintfH("interface %lu alt %lu: bInterfaceNumber=%lu bAlternateSetting=%lu bNumEndpoints=%lu bInterfaceClass=0x%02lx bInterfaceSubClass=0x%02lx bInterfaceProtocol=0x%02lx iInterface=%lu\n",
-                     (ULONG)if_index,
-                     (ULONG)current_alt_index,
-                     (ULONG)ifd->bInterfaceNumber,
-                     (ULONG)ifd->bAlternateSetting,
-                     (ULONG)ifd->bNumEndpoints,
-                     (ULONG)ifd->bInterfaceClass,
-                     (ULONG)ifd->bInterfaceSubClass,
-                     (ULONG)ifd->bInterfaceProtocol,
-                     (ULONG)ifd->iInterface);
-
-            if (current_if->active_altsetting == NULL || current_alt->desc.bAlternateSetting == 0)
-                current_if->active_altsetting = current_alt;
-            break;
-        }
-        case USB_DT_ENDPOINT:
-        {
-            if (!current_if || !current_alt)
-            {
-                Kprintf("endpoint without interface or altsetting\n");
-                break;
-            }
-            if (current_alt->no_of_ep >= USB_MAXENDPOINTS)
-            {
-                Kprintf("too many endpoints for interface %lu alt %lu\n",
-                        (ULONG)if_index, (ULONG)current_alt_index);
-                break;
-            }
-
-            struct usb_endpoint_descriptor *epd = (struct usb_endpoint_descriptor *)cursor;
-            u32 ep_idx = current_alt->no_of_ep;
-            CopyMem(epd, &current_alt->ep_desc[ep_idx], sizeof(struct usb_endpoint_descriptor));
-            KprintfH("  endpoint %lu: bEndpointAddress=0x%02lx bmAttributes=0x%02lx wMaxPacketSize=%lu bInterval=%lu\n",
-                     (ULONG)ep_idx,
-                     (ULONG)epd->bEndpointAddress,
-                     (ULONG)epd->bmAttributes,
-                     (ULONG)le16(epd->wMaxPacketSize),
-                     (ULONG)epd->bInterval);
-
-            current_alt->no_of_ep++;
-            break;
-        }
-        case USB_DT_SS_ENDPOINT_COMP:
-        {
-            KprintfH("found SS EP COMP descriptor\n");
-            if (current_if && current_alt && current_alt->no_of_ep > 0)
-            {
-                struct usb_ss_ep_comp_descriptor *comp = (struct usb_ss_ep_comp_descriptor *)cursor;
-                u32 ep_slot = (u32)(current_alt->no_of_ep - 1U);
-                CopyMem(comp, &current_alt->ss_ep_comp_desc[ep_slot], sizeof(struct usb_ss_ep_comp_descriptor));
-            }
-            break;
-        }
-        default:
-            // Skip class- or vendor-specific descriptors gracefully.
-            KprintfH("found class/vendor-specific descriptor 0x%lx, len=%lu\n", (ULONG)dtype, (ULONG)dlen);
-            break;
-        }
-
-        cursor += dlen;
-    }
-    KprintfH("parsed config with %lu interfaces\n", (ULONG)conf->no_of_if);
-
-    if (conf->no_of_if != desc->bNumInterfaces)
-    {
-        Kprintf("interface count mismatch %lu != %lu\n",
-                (ULONG)conf->no_of_if, (ULONG)desc->bNumInterfaces);
-        goto error;
-    }
-
-    for (struct MinNode *n = udev->configurations.mlh_Head; n->mln_Succ; n = n->mln_Succ)
-    {
-        struct usb_config *oldconf = (struct usb_config *)n;
-        if (oldconf->desc.bConfigurationValue == conf->desc.bConfigurationValue)
-        {
-            KprintfH("removing old config with value %lu\n", (ULONG)oldconf->desc.bConfigurationValue);
-            RemoveMinNode(n);
-            pool_free(udev->controller->metaPool, oldconf);
-            break;
-        }
-    }
-    AddHeadMinList(&udev->configurations, (struct MinNode *)conf);
-
-    return;
-
-error:
-    pool_free(udev->controller->metaPool, conf);
-}
-
-static void xhci_filter_ss_ep_companion_desc(struct USBIORequest *io)
-{
-    if (!io->data_buffer || io->actual_length < sizeof(struct usb_config_descriptor))
-        return;
-
-    struct usb_config_descriptor *desc = (struct usb_config_descriptor *)io->data_buffer;
-    if (desc->bDescriptorType != USB_DT_CONFIG)
-        return;
-
-    u16 total_len = le16(desc->wTotalLength);
-    if (total_len > io->actual_length)
-        total_len = (u16)io->actual_length;
-
-    u8 *read = io->data_buffer + desc->bLength;
-    u8 *write = read;
-    u8 *end = io->data_buffer + total_len;
-
-    while (read + 2 <= end)
-    {
-        u8 dlen = read[0];
-        u8 dtype = read[1];
-        if (dlen == 0 || read + dlen > end)
-            break;
-
-        if (dtype != USB_DT_SS_ENDPOINT_COMP)
-        {
-            if (write != read)
-            {
-                for (u8 i = 0; i < dlen; ++i)
-                    write[i] = read[i];
-            }
-            write += dlen;
-        }
-
-        read += dlen;
-    }
-
-    if (write < end)
-        mem_zero(write, (ULONG)(end - write));
-
-    u16 new_total = (u16)(write - (u8 *)io->data_buffer);
-    if (new_total != total_len)
-        desc->wTotalLength = le16(new_total);
-
-    io->actual_length = new_total;
-}
-
 static BOOL xhci_udev_iface_has_active_rt_iso(struct usb_device *udev, u8 iface_number)
 {
     if (!udev || !udev->active_config)
@@ -1745,7 +1011,7 @@ static BOOL xhci_udev_iface_has_active_rt_iso(struct usb_device *udev, u8 iface_
     return FALSE;
 }
 
-static void xhci_udev_disconnect(struct usb_device *udev, BOOL recursive)
+void xhci_udev_disconnect(struct usb_device *udev, BOOL recursive)
 {
     if (!udev || !udev->slot_id)
         return;
@@ -1784,45 +1050,7 @@ static void xhci_udev_disconnect(struct usb_device *udev, BOOL recursive)
     xhci_disable_slot(udev);
 }
 
-static enum usb_device_speed xhci_udev_speed_from_port_status(u16 status)
-{
-    switch (status & USB_PORT_STAT_SPEED_MASK)
-    {
-    case USB_PORT_STAT_HIGH_SPEED:
-        return USB_SPEED_HIGH;
-    case USB_PORT_STAT_LOW_SPEED:
-        return USB_SPEED_LOW;
-    default:
-        return USB_SPEED_FULL;
-    }
-}
-
-static enum usb_device_speed xhci_udev_speed_from_ss_port_status(u16 status)
-{
-    switch (status & USB_SS_PORT_STAT_SPEED)
-    {
-    case USB_SS_PORT_STAT_SPEED_LOW:
-        return USB_SPEED_LOW;
-    case USB_SS_PORT_STAT_SPEED_FULL:
-        return USB_SPEED_FULL;
-    case USB_SS_PORT_STAT_SPEED_HIGH:
-        return USB_SPEED_HIGH;
-    case USB_SS_PORT_STAT_SPEED_5GBPS:
-        return USB_SPEED_SUPER;
-    default:
-        return USB_SPEED_UNKNOWN;
-    }
-}
-
-static BOOL xhci_udev_ss_port_ready_for_attach(u16 status, enum usb_device_speed speed)
-{
-    return (status & USB_PORT_STAT_CONNECTION) != 0 &&
-           (status & USB_PORT_STAT_ENABLE) != 0 &&
-           (status & USB_PORT_STAT_RESET) == 0 &&
-           speed != USB_SPEED_UNKNOWN;
-}
-
-static struct usb_device *xhci_udev_find_child_on_port(struct usb_device *hub, u32 port)
+struct usb_device *xhci_udev_find_child_on_port(struct usb_device *hub, u32 port)
 {
     if (!hub)
         return NULL;
@@ -1839,150 +1067,6 @@ static struct usb_device *xhci_udev_find_child_on_port(struct usb_device *hub, u
     }
 
     return NULL;
-}
-
-static void xhci_udev_cache_ss_hub_descriptor(struct usb_device *udev, struct usb_hub_descriptor *hub, u32 actual)
-{
-    if (!udev || !hub || actual < 4)
-        return;
-
-    u8 len = hub->bLength;
-    if (len == 0 || len > actual)
-        len = (u8)(actual < sizeof(struct usb_hub_descriptor) ? actual : sizeof(struct usb_hub_descriptor));
-
-    CopyMem(hub, &udev->ss_hub_desc, len);
-    KprintfH("Cached SS hub descriptor for addr %lu with %lu ports\n",
-             (ULONG)udev->virtual_address, (ULONG)hub->bNbrPorts);
-}
-
-static void xhci_udev_set_ss_hub_depth(struct usb_device *udev)
-{
-    if (!udev || !udev->is_hub || !udev->ss_hub_emulation || udev->speed < USB_SPEED_SUPER)
-        return;
-
-    /* External hub only: root hub does not need this request. */
-    if (!udev->parent)
-        return;
-
-    if (udev->ss_hub_depth_set)
-        return;
-
-    KprintfH("SS hub addr=%lu route=0x%lx -> SET_HUB_DEPTH depth=%lu\n",
-             (ULONG)udev->virtual_address, (ULONG)udev->route, (ULONG)udev->route_depth);
-
-    xhci_udev_send_control_request(udev,
-                                   0,
-                                   USB_DIR_OUT | USB_RT_HUB,
-                                   USB_REQ_SET_HUB_DEPTH,
-                                   udev->route_depth /* wValue */,
-                                   0 /* wIndex */,
-                                   0 /* wLength */,
-                                   FALSE /* enqueue */);
-
-    udev->ss_hub_depth_set = TRUE;
-}
-
-static u32 xhci_udev_build_usb2_hub_descriptor(struct usb_device *udev, u8 *buf, const u32 max_len)
-{
-    if (!udev || !buf || max_len == 0)
-        return 0;
-
-    struct usb_hub_descriptor hub;
-    mem_zero(&hub, sizeof(hub));
-
-    const u8 ports = udev->hub_num_ports;
-    u32 needed_words = ((u32)ports + 1U + 7U) / 8U;
-    const u8 needed = (u8)(needed_words < sizeof(hub.u.hs.DeviceRemovable) ? needed_words : sizeof(hub.u.hs.DeviceRemovable));
-
-    hub.bLength = (u8)(7U + 2U * needed);
-    hub.bDescriptorType = USB_DT_HUB;
-    hub.bNbrPorts = ports;
-
-    hub.wHubCharacteristics = udev->ss_hub_desc.wHubCharacteristics;
-    hub.bPwrOn2PwrGood = udev->ss_hub_desc.bPwrOn2PwrGood;
-    hub.bHubContrCurrent = udev->ss_hub_desc.bHubContrCurrent;
-
-    for (u8 i = 0; i < needed; ++i)
-        hub.u.hs.PortPowerCtrlMask[i] = 0xFF;
-
-    u32 actual = max_len < hub.bLength ? max_len : hub.bLength;
-    CopyMem(&hub, buf, actual);
-    return actual;
-}
-
-/* Translate SS hub descriptor request: modify request to ask for SS descriptor,
- * it will be translated back to USB 2.0 in the parse handler */
-static void xhci_udev_translate_hub_descriptor_request(struct usb_device *udev, struct USBIORequest *io)
-{
-    if (!udev->ss_hub_emulation || !io || !io->data_buffer || io->data_buffer_length == 0)
-        return;
-
-    struct USBSetupPacket *setup = &io->setup;
-    const u8 descriptorType = (le16(setup->wValue) >> 8) & 0xFFU;
-    const u16 typeReq = (u16)(((u16)setup->bmRequestType << 8) | setup->bRequest);
-
-    /* Only translate GetHubDescriptor requests for USB_DT_HUB */
-    if (typeReq != GetHubDescriptor || descriptorType != USB_DT_HUB)
-        return;
-
-    /* Modify the request to ask for SS hub descriptor instead */
-    u16 old_value = le16(setup->wValue);
-    setup->wValue = le16((USB_DT_SS_HUB << 8) | (old_value & 0xFF));
-
-    KprintfH("SS hub addr=%lu: modified wValue from 0x%04lx (USB_DT_HUB) to 0x%04lx (USB_DT_SS_HUB)\n",
-             (ULONG)udev->virtual_address, (ULONG)old_value, (ULONG)le16(setup->wValue));
-
-    /* Return FALSE to let the request proceed normally - it will be translated back in parse */
-    return;
-}
-
-static void xhci_udev_map_ss_port_status(u16 *wStatus, u16 *wChange, enum usb_device_speed speed)
-{
-    u16 wStatusNew = *wStatus & USB_SS_PORT_STAT_MASK;
-
-    if ((*wStatus & PORT_PLS_MASK) == XDEV_U3)
-    {
-        KprintfH("SS hub: PLS=U3 detected, mapping to USB_PORT_STAT_SUSPEND\n");
-        wStatusNew |= USB_PORT_STAT_SUSPEND;
-    }
-    if (*wStatus & USB_SS_PORT_STAT_POWER)
-    {
-        KprintfH("SS hub: POWER bit set, mapping to USB_PORT_STAT_POWER\n");
-        wStatusNew |= USB_PORT_STAT_POWER;
-    }
-
-    switch (speed)
-    {
-    case USB_SPEED_LOW:
-        KprintfH("SS hub: detected LowSpeed device, mapping to USB_PORT_STAT_LOW_SPEED\n");
-        wStatusNew |= USB_PORT_STAT_LOW_SPEED;
-        break;
-    case USB_SPEED_FULL:
-        KprintfH("SS hub: detected FullSpeed device\n");
-        break;
-    case USB_SPEED_HIGH:
-        KprintfH("SS hub: detected HighSpeed device, mapping to USB_PORT_STAT_HIGH_SPEED\n");
-        wStatusNew |= USB_PORT_STAT_HIGH_SPEED;
-        break;
-    default:
-        KprintfH("SS hub: detected SuperSpeed device, mapping to USB_PORT_STAT_HIGH_SPEED for compatibility\n");
-        wStatusNew |= USB_PORT_STAT_HIGH_SPEED;
-        break;
-    }
-
-    KprintfH("SS hub: mapped status 0x%04lx -> 0x%04lx\n", (ULONG)*wStatus, (ULONG)wStatusNew);
-
-    u16 wChangeNew = *wChange & (USB_PORT_STAT_C_CONNECTION | USB_PORT_STAT_C_OVERCURRENT | USB_PORT_STAT_C_RESET);
-
-    if (*wChange & USB_SS_PORT_STAT_C_LINK_STATE && ((*wStatus & PORT_PLS_MASK) == XDEV_U0))
-    {
-        KprintfH("SS hub: C_LINK_STATE detected and PLS=U0, mapping to C_SUSPEND\n");
-        wChangeNew |= USB_PORT_STAT_C_SUSPEND;
-    }
-
-    KprintfH("SS hub: mapped change 0x%04lx -> 0x%04lx\n", (ULONG)*wChange, (ULONG)wChangeNew);
-    *wStatus = wStatusNew;
-    *wChange = wChangeNew;
 }
 
 static void handle_get_device_descriptor(struct usb_device *udev, struct USBIORequest *io)
@@ -2038,48 +1122,6 @@ static void handle_get_device_descriptor(struct usb_device *udev, struct USBIORe
     }
 }
 
-static void handle_get_hub_descriptor(struct usb_device *udev, struct USBIORequest *io, u8 descriptorType)
-{
-    if (!io->data_buffer || io->actual_length < 5)
-        return;
-
-    struct usb_hub_descriptor *hub = (struct usb_hub_descriptor *)io->data_buffer;
-    KprintfH("Hub Descriptor: bLength=%lu bDescriptorType=%lu bNbrPorts=%lu wHubCharacteristics=0x%04lx bPwrOn2PwrGood=%lu bHubContrCurrent=%lu\n",
-             (ULONG)hub->bLength,
-             (ULONG)hub->bDescriptorType,
-             (ULONG)hub->bNbrPorts,
-             (ULONG)le16(hub->wHubCharacteristics),
-             (ULONG)hub->bPwrOn2PwrGood,
-             (ULONG)hub->bHubContrCurrent);
-
-    /* Update TT think time if changed */
-    if (udev->parent)
-    {
-        const u16 characteristics = le16(hub->wHubCharacteristics);
-        udev->tt_think_time = (u8)((characteristics >> 5) & 0x3);
-        KprintfH("hub addr %lu TT think time code=%lu (bit-times=%lu)\n",
-                 (ULONG)udev->virtual_address, (ULONG)udev->tt_think_time, (ULONG)((udev->tt_think_time + 1) * 8));
-    }
-
-    udev->hub_num_ports = hub->bNbrPorts;
-
-    /* If this is an SS hub descriptor response, cache it */
-    if (descriptorType == USB_DT_SS_HUB)
-    {
-        KprintfH("SS hub addr=%lu: caching USB3 hub descriptor (len=%lu)\n", (ULONG)udev->virtual_address, (ULONG)io->actual_length);
-        xhci_udev_cache_ss_hub_descriptor(udev, hub, io->actual_length);
-        xhci_udev_set_ss_hub_depth(udev);
-
-        /* If the stack requested USB 2.0 descriptor but we fetched SS, translate it */
-        if (udev->ss_hub_emulation)
-        {
-            /* Build USB 2.0 descriptor from the SS descriptor we just cached */
-            io->actual_length = xhci_udev_build_usb2_hub_descriptor(udev, (u8 *)io->data_buffer, io->data_buffer_length);
-            KprintfH("SS hub addr=%lu: translated USB3 descriptor to USB2 format. Size %lu bytes\n", (ULONG)udev->virtual_address, (ULONG)io->actual_length);
-        }
-    }
-}
-
 static void xhci_trim_string_descriptor(struct USBIORequest *io)
 {
     if (!io || !io->data_buffer || io->actual_length < 2)
@@ -2129,83 +1171,6 @@ static void xhci_append_ss_suffix(struct usb_device *udev, struct USBIORequest *
 
     str_desc->bLength = (u8)(str_desc->bLength + (u8)suffix_len);
     io->actual_length = (u32)str_desc->bLength < io->data_buffer_length ? (u32)str_desc->bLength : io->data_buffer_length;
-}
-
-static void handle_get_port_status(struct usb_device *udev, struct USBIORequest *io)
-{
-    if (!io->data_buffer || io->actual_length < 4)
-        return;
-
-    struct xhci_ctrl *ctrl = udev->controller;
-    if (!ctrl)
-        return;
-    const u8 port = le16(io->setup.wIndex) & 0xFFu;
-
-    u16 wStatus = le16(((u16 *)io->data_buffer)[0]);
-    u16 wChange = le16(((u16 *)io->data_buffer)[1]);
-    const u16 rawStatus = wStatus;
-    const u16 rawChange = wChange;
-    /* Extract speed from the appropriate bit positions based on hub type */
-    enum usb_device_speed speed = (udev->ss_hub_emulation) ? xhci_udev_speed_from_ss_port_status(rawStatus) : xhci_udev_speed_from_port_status(wStatus);
-
-    if (udev->ss_hub_emulation)
-    {
-        xhci_udev_map_ss_port_status(&wStatus, &wChange, speed);
-        ((u16 *)io->data_buffer)[0] = le16(wStatus);
-        ((u16 *)io->data_buffer)[1] = le16(wChange);
-    }
-
-    KprintfH("hub addr=%lu port=%lu status=%04lx change=%04lx\n", (ULONG)udev->virtual_address, (ULONG)port, (ULONG)wStatus, (ULONG)wChange);
-
-    /* Tear down any existing child as soon as the port is powered-but-disabled,
-     * otherwise re-enumeration races the stale slot/context we still own. */
-    const BOOL port_lost_child = ((wStatus & USB_PORT_STAT_POWER) == 0) ||
-                                 ((wStatus & USB_PORT_STAT_CONNECTION) == 0) ||
-                                 ((wStatus & USB_PORT_STAT_CONNECTION) != 0 &&
-                                  (wStatus & USB_PORT_STAT_ENABLE) == 0 &&
-                                  (wStatus & USB_PORT_STAT_RESET) == 0);
-
-    if (port_lost_child)
-    {
-        KprintfH("hub addr=%lu port=%lu lost power, disconnected, or disabled; removing child if any\n",
-                 (ULONG)udev->virtual_address, (ULONG)port);
-        struct usb_device *child = xhci_udev_find_child_on_port(udev, port);
-        if (child)
-        {
-            KprintfH("hub addr=%lu port=%lu tearing down child addr=%lu slot=%lu before re-enumeration\n",
-                     (ULONG)udev->virtual_address, (ULONG)port, (ULONG)child->virtual_address, (ULONG)child->slot_id);
-            xhci_udev_disconnect(child, TRUE);
-        }
-    }
-
-    /* SS hub: use raw (pre-mapping) status to detect attach readiness.
-     * USB 3.0 ports transition to enabled automatically after link training,
-     * so we wait for connected + enabled + known speed before arming
-     * pending_parent for the next SET_ADDRESS. */
-    if (udev->ss_hub_emulation)
-    {
-        if (xhci_udev_ss_port_ready_for_attach(rawStatus, speed) &&
-            (rawChange & (USB_PORT_STAT_C_CONNECTION |
-                          USB_PORT_STAT_C_RESET |
-                          USB_SS_PORT_STAT_C_BH_RESET |
-                          USB_SS_PORT_STAT_C_LINK_STATE)))
-        {
-            KprintfH("hub addr=%lu port=%lu speed=%lu SS attach ready; remembering for pending attach (raw_status=%04lx)\n",
-                     (ULONG)udev->virtual_address, (ULONG)port, (ULONG)speed, (ULONG)rawStatus);
-            ctrl->pending_parent = udev;
-            ctrl->pending_parent_port = port;
-            ctrl->pending_parent_speed = speed;
-        }
-    }
-    /* USB 2.0 enables device after reset completes */
-    else if ((wChange & USB_PORT_STAT_C_RESET) && (wStatus & (USB_PORT_STAT_CONNECTION | USB_PORT_STAT_ENABLE)))
-    {
-        KprintfH("hub addr=%lu port=%lu speed=%lu reset-complete; remembering for pending attach (status=%04lx)\n",
-                 (ULONG)udev->virtual_address, (ULONG)port, (ULONG)speed, (ULONG)wStatus);
-        ctrl->pending_parent = udev;
-        ctrl->pending_parent_port = port;
-        ctrl->pending_parent_speed = speed;
-    }
 }
 
 static void handle_set_address(struct usb_device *udev, struct USBIORequest *io)
@@ -2284,10 +1249,10 @@ static void xhci_udev_parse_control_message(struct usb_device *udev, struct USBI
             /* If this was a successful GET_DESCRIPTOR(CONFIGURATION),
              * cache the configuration descriptor for later use.
              */
-            parse_config_descriptor(udev, (u8 *)io->data_buffer, (u16)io->actual_length);
+            xhci_parse_config_descriptor(udev, (u8 *)io->data_buffer, (u16)io->actual_length);
 
             /* USB 2.0 stacks  don't like seeing SS companion descriptors */
-            xhci_filter_ss_ep_companion_desc(io);
+            xhci_hub_filter_ss_ep_companion_desc(io);
             break;
         case USB_DT_DEVICE:
             /* Update FS control endpoint max packet size based on device descriptor. */
@@ -2308,12 +1273,12 @@ static void xhci_udev_parse_control_message(struct usb_device *udev, struct USBI
     case GetHubDescriptor:
         /* Record TT think time from hub descriptors so child devices can be programmed correctly. */
         if (descriptorType == USB_DT_HUB || descriptorType == USB_DT_SS_HUB)
-            handle_get_hub_descriptor(udev, io, descriptorType);
+            xhci_hub_handle_get_descriptor(udev, io, descriptorType);
         break;
 
     case GetPortStatus:
         /* Detect downstream port disconnects via hub GET_STATUS replies. */
-        handle_get_port_status(udev, io);
+        xhci_hub_handle_get_port_status(udev, io);
         break;
 
     case ClearPortFeature:
@@ -2345,5 +1310,247 @@ static void xhci_udev_parse_control_message(struct usb_device *udev, struct USBI
         if (le16(io->setup.wValue) != 0)
             xhci_lpm_enable(udev);
         break;
+    }
+}
+
+void xhci_udev_io_reply_data(struct usb_device *udev, struct USBIORequest *io, s8 err, u32 actual)
+{
+    if (!io || !udev)
+        return;
+
+    io->actual_length = actual;
+    io->req.io_Error = err;
+
+    if (io->req.io_Command == CMD_REQUEST_CONTROL && err == ERR_NO_ERROR && io->endpoint == 0)
+        xhci_udev_parse_control_message(udev, io);
+
+    KprintfH("err=%ld actual=%lu\n", (LONG)err, (ULONG)actual);
+
+    /* Internal, reply-less requests (IOF_QUICK + magic tag) */
+    if (io->driver_private_flags & REQ_INTERNAL)
+    {
+        xhci_udev_complete_internal(udev, io);
+        return;
+    }
+
+    ReplyMsg((struct Message *)io);
+}
+
+/* ---- Request dispatch entry points -------------------------------------- */
+
+/* SET_CONFIGURATION special case: stash the request as UDEV_OP_CONFIGURE, then
+ * optionally pre-fetch a descriptor on EP0 before configuring; the prefetch
+ * completion resumes the deferred SET_CONFIGURATION
+ * (xhci_udev_run_pending_set_config).
+ *   - Hubs: GET_DESCRIPTOR(hub) so xhci_set_configuration() can program correct
+ *     Number of Ports / TT Think Time into the slot context.
+ *   - HS/SS: GET_DESCRIPTOR(BOS) for U1/U2 exit latencies (USB3) or BESL (USB2
+ *     LPM), needed for Max Exit Latency.  Non-fatal.
+ * EP0 is live (set up during ADDRESS_DEVICE).  A prefetch that can't be
+ * submitted leaves the stash intact, so we just run the config now.
+ * Returns TRUE when this was a SET_CONFIGURATION (always fully handled). */
+static BOOL xhci_udev_ctrl_set_config(struct usb_device *udev, struct USBIORequest *io)
+{
+    struct USBSetupPacket *setup = &io->setup;
+    if (!(setup->bRequest == USB_REQ_SET_CONFIGURATION &&
+          (setup->bmRequestType & USB_TYPE_MASK) == USB_TYPE_STANDARD))
+        return FALSE;
+
+    if (!xhci_udev_op_begin(udev, UDEV_OP_CONFIGURE, io))
+    {
+        /* op slot busy (unexpected): configure without the prefetch */
+        xhci_udev_do_set_config(udev, io);
+        return TRUE;
+    }
+
+    if (udev->is_hub)
+    {
+        if (!udev->hub_desc_fetched && xhci_udev_fetch_hub_descriptor(udev))
+            return TRUE;
+    }
+    else if (udev->speed >= USB_SPEED_HIGH)
+    {
+        if (!udev->bos_fetched && xhci_udev_fetch_bos(udev))
+            return TRUE;
+    }
+
+    /* LS/FS, or the prefetch wasn't submitted: configure immediately. */
+    xhci_udev_run_pending_set_config(udev);
+    return TRUE;
+}
+
+/* CLEAR_FEATURE(ENDPOINT_HALT) from the stack: keep the controller-side
+ * endpoint state in sync with the device (mirror Linux xhci_endpoint_reset).
+ * If our own STALL recovery already reset the endpoint and sent the
+ * device-level clear-halt, answer without a duplicate wire request; if the
+ * endpoint is still halted (the stack noticed before our event handling), run
+ * our recovery - it ends with the internal clear-halt.  Returns TRUE only when
+ * the request was answered here; an unsolicited clear on a healthy endpoint
+ * (e.g. mass-storage reset recovery) returns FALSE and is forwarded as-is: the
+ * device resets its data toggle/sequence, and a stale xHC sequence recovers
+ * through the normal STALL path on the next transfer. */
+static BOOL xhci_udev_ctrl_clear_ep_halt(struct usb_device *udev, struct USBIORequest *io)
+{
+    struct USBSetupPacket *setup = &io->setup;
+    if (!(setup->bmRequestType == (USB_DIR_OUT | USB_TYPE_STANDARD | USB_RECIP_ENDPOINT) &&
+          setup->bRequest == USB_REQ_CLEAR_FEATURE &&
+          le16(setup->wValue) == USB_ENDPOINT_HALT &&
+          udev->slot_id != 0))
+        return FALSE;
+
+    u8 ep_addr = (u8)(le16(setup->wIndex) & 0xffU);
+    u8 halt_ep_index = xhci_ep_index_from_parts(ep_addr & 0x0fU,
+                                                (ep_addr & USB_DIR_IN) ? DIRECTION_IN : DIRECTION_OUT);
+    struct ep_context *halt_ep_ctx = (halt_ep_index != 0)
+                                         ? xhci_ep_get_context_for_index(udev, halt_ep_index)
+                                         : NULL;
+    if (!halt_ep_ctx)
+        return FALSE;
+
+    BOOL synced = xhci_ep_consume_halt_synced(halt_ep_ctx);
+    if (!synced && xhci_read_hw_ep_state(udev, halt_ep_index) == EP_STATE_HALTED)
+    {
+        KprintfH("stack clear-halt on halted EP %lu: running recovery\n", (ULONG)halt_ep_index);
+        xhci_reset_ep(udev, halt_ep_index);
+        synced = TRUE;
+    }
+    if (!synced)
+        return FALSE;
+
+    io->req.io_Error = ERR_NO_ERROR;
+    io->actual_length = 0;
+    if (!(io->req.io_Flags & IOF_QUICK))
+        ReplyMsg((struct Message *)io);
+    return TRUE;
+}
+
+/* SetPortFeature(SUSPEND) to an external hub: stop the attached child's
+ * endpoint rings first (xHCI 4.15.1); the request is forwarded to the hub once
+ * the stops complete (xhci_udev_suspend_finish).  Returns TRUE if the child's
+ * stop sequence was started (request deferred), FALSE to forward immediately. */
+static BOOL xhci_udev_ctrl_hub_suspend(struct usb_device *udev, struct USBIORequest *io)
+{
+    struct USBSetupPacket *setup = &io->setup;
+    if (!(udev->is_hub &&
+          setup->bmRequestType == (USB_DIR_OUT | USB_RT_PORT) &&
+          setup->bRequest == USB_REQ_SET_FEATURE &&
+          le16(setup->wValue) == USB_PORT_FEAT_SUSPEND))
+        return FALSE;
+
+    struct usb_device *child = xhci_udev_find_child_on_port(udev, le16(setup->wIndex) & 0xffU);
+    return (child && xhci_udev_suspend_device(child, 0, io)) ? TRUE : FALSE;
+}
+
+static s8 xhci_udev_send_ctrl_first(struct usb_device *udev, struct USBIORequest *io, u32 timeout_ms)
+{
+    /* Work around class drivers that omit the direction bit in endpoint-recipient requests (e.g., UAC1 SET_CUR). */
+    xhci_udev_patch_endpoint_address(udev, io);
+
+    KprintfH("bmReqType=%02lx bReq=%02lx wValue=%04lx wIndex=%04lx wLength=%04lx\n",
+             (ULONG)io->setup.bmRequestType,
+             (ULONG)io->setup.bRequest,
+             le16(io->setup.wValue),
+             le16(io->setup.wIndex),
+             le16(io->setup.wLength));
+
+    /* Translate USB 2.0 hub requests to SS format for SS hubs */
+    xhci_hub_translate_descriptor_request(udev, io);
+
+    if (xhci_hub_filter_emulated_ctrl_request(udev, io))
+        return ERR_NO_ERROR;
+
+    struct xhci_ctrl *ctrl = udev->controller;
+    if (io->virtual_address == xhci_roothub_get_address(ctrl->root_hub))
+    {
+        xhci_roothub_submit_ctrl_request(ctrl->root_hub, io);
+        xhci_udev_parse_control_message(udev, io);
+        if (io->req.io_Error != ERR_NO_ERROR)
+            return io->req.io_Error;
+        if (!(io->req.io_Flags & IOF_QUICK))
+            ReplyMsg((struct Message *)io);
+        return ERR_NO_ERROR;
+    }
+
+    struct USBSetupPacket *setup = &io->setup;
+    if (setup->bRequest == USB_REQ_SET_ADDRESS && (setup->bmRequestType & USB_TYPE_MASK) == USB_TYPE_STANDARD)
+    {
+        xhci_address_device(udev, io);
+        return ERR_NO_ERROR;
+    }
+
+    /* Standard-request special cases, each fully handled when it applies. */
+    if (xhci_udev_ctrl_set_config(udev, io))
+        return ERR_NO_ERROR;
+
+    if (xhci_udev_ctrl_clear_ep_halt(udev, io))
+        return ERR_NO_ERROR;
+
+    if (xhci_udev_ctrl_hub_suspend(udev, io))
+        return ERR_NO_ERROR;
+
+    /* If we don't have a slot yet, enable one and allocate Virt Dev */
+    if (udev->slot_id == 0 && udev->virtual_address == 0)
+    {
+        // this will store the req and submit it once addressed
+        xhci_address_device(udev, io);
+        return ERR_NO_ERROR;
+    }
+
+    return xhci_ring_enqueue_td(udev, io, timeout_ms, FALSE);
+}
+
+s8 xhci_udev_send(struct USBIORequest *req)
+{
+    struct XHCIUnit *unit = (struct XHCIUnit *)req->req.io_Unit;
+    if (!unit)
+    {
+        Kprintf("missing unit pointer (cmd=%lu, req=%lx, devaddr=%lu)\n",
+                (ULONG)req->req.io_Command, (ULONG)req, (ULONG)req->virtual_address);
+        return ERR_BAD_PARAMETERS;
+    }
+
+    struct usb_device *udev = xhci_udev_get(unit, req->virtual_address);
+    if (!udev)
+    {
+        KprintfH("Device does not exist for addr %lu\n", (ULONG)req->virtual_address);
+        return ERR_TIMEOUT;
+    }
+
+    u32 timeout_ms = 0;
+    if ((req->flags & DRIVER_FLAG_TIMEOUT_DEFINED))
+        timeout_ms = req->timeout;
+
+    KprintfH("dev=%lx addr=%lu slot=%lu ep=%lu dir=%s len=%lu flags=%lx tmo=%lu\n",
+             (ULONG)udev, (ULONG)udev->virtual_address, (ULONG)udev->slot_id,
+             (ULONG)(req->endpoint & 0x0F), (req->direction == DIRECTION_IN) ? "IN" : "OUT",
+             (ULONG)req->data_buffer_length, (ULONG)req->flags,
+             (ULONG)timeout_ms);
+
+    switch (req->req.io_Command)
+    {
+    case CMD_REQUEST_CONTROL:
+        return xhci_udev_send_ctrl_first(udev, req, timeout_ms);
+    case CMD_REQUEST_ISOCHRONOUS:
+    case CMD_REQUEST_BULK:
+        return xhci_ring_enqueue_td(udev, req, timeout_ms, FALSE);
+    case CMD_REQUEST_INTERRUPT:
+    {
+        struct xhci_ctrl *ctrl = unit->xhci_ctrl;
+        if (udev->virtual_address == xhci_roothub_get_address(ctrl->root_hub))
+        {
+            s8 result = xhci_roothub_submit_int_request(ctrl->root_hub, req);
+            return result;
+        }
+
+        return xhci_ring_enqueue_td(udev, req, timeout_ms, FALSE);
+    }
+    default:
+        Kprintf("unsupported command %lu (req=%lx, devaddr=%lu, endpoint=%lu, flags=0x%lx)\n",
+                (ULONG)req->req.io_Command,
+                (ULONG)req,
+                (ULONG)req->virtual_address,
+                (ULONG)req->endpoint,
+                (ULONG)req->flags);
+        return ERR_BAD_PARAMETERS;
     }
 }
